Move WorkoutSession to backend
