+++ tests/test_workout_session.py
@@ -1,4 +1,5 @@
 import core
+from backend.workout_session import WorkoutSession
 import sqlite3
 import pytest
 
@@ -7,7 +8,7 @@ def test_workout_session_flow(sample_db):
     presets = core.load_workout_presets(sample_db)
     assert any(p["name"] == "Push Day" for p in presets)
 
-    session = core.WorkoutSession("Push Day", db_path=sample_db, rest_duration=1)
+    session = WorkoutSession("Push Day", db_path=sample_db, rest_duration=1)
     assert session.next_exercise_display() == "Push-up set 1 of 2"
     assert session.upcoming_exercise_display() == "Push-up set 2 of 2"
 
@@ -32,7 +33,7 @@ def test_workout_session_flow(sample_db):
 
 
 def test_pre_set_metrics_flow(sample_db):
-    session = core.WorkoutSession("Push Day", db_path=sample_db, rest_duration=1)
+    session = WorkoutSession("Push Day", db_path=sample_db, rest_duration=1)
 
     # complete push-up sets to reach Bench Press
     session.record_metrics(session.current_exercise, session.current_set, {"Reps": 10})
@@ -54,7 +55,7 @@ def test_pre_set_metrics_flow(sample_db):
 
 
 def test_pre_set_metrics_require_selection(sample_db):
-    session = core.WorkoutSession("Push Day", db_path=sample_db, rest_duration=1)
+    session = WorkoutSession("Push Day", db_path=sample_db, rest_duration=1)
 
     session.record_metrics(session.current_exercise, session.current_set, {"Reps": 10})
     session.mark_set_completed()
@@ -79,7 +80,7 @@ def test_rest_time_uses_next_exercise(sample_db):
     conn.commit()
     conn.close()
 
-    session = core.WorkoutSession("Push Day", db_path=sample_db)
+    session = WorkoutSession("Push Day", db_path=sample_db)
     assert session.rest_duration == 10
 
     session.record_metrics(session.current_exercise, session.current_set, {"Reps": 10})
@@ -92,7 +93,7 @@ def test_rest_time_uses_next_exercise(sample_db):
 
 
 def test_required_post_set_metrics(sample_db):
-    session = core.WorkoutSession("Push Day", db_path=sample_db, rest_duration=1)
+    session = WorkoutSession("Push Day", db_path=sample_db, rest_duration=1)
     session.mark_set_completed()
     assert not session.has_required_post_set_metrics()
     session.record_metrics(session.current_exercise, session.current_set, {"Reps": 10})
@@ -100,7 +101,7 @@ def test_required_post_set_metrics(sample_db):
 
 
 def test_mark_set_completed_time_adjustment(sample_db, monkeypatch):
-    session = core.WorkoutSession("Push Day", db_path=sample_db, rest_duration=30)
+    session = WorkoutSession("Push Day", db_path=sample_db, rest_duration=30)
     session.exercises[0]["rest"] = 30
     session.record_metrics(session.current_exercise, session.current_set, {"Reps": 10})
     monkeypatch.setattr(core.time, "time", lambda: 165.0)
@@ -110,7 +111,7 @@ def test_mark_set_completed_time_adjustment(sample_db, monkeypatch):
 
 
 def test_undo_last_set_restores_metrics_and_timer(sample_db, monkeypatch):
-    session = core.WorkoutSession("Push Day", db_path=sample_db, rest_duration=1)
+    session = WorkoutSession("Push Day", db_path=sample_db, rest_duration=1)
     start = session.current_set_start_time
 
     monkeypatch.setattr(core.time, "time", lambda: start + 5)
@@ -128,7 +129,7 @@ def test_undo_last_set_restores_metrics_and_timer(sample_db, monkeypatch):
 
 
 def test_undo_set_start_returns_to_rest(sample_db, monkeypatch):
-    session = core.WorkoutSession("Push Day", db_path=sample_db, rest_duration=10)
+    session = WorkoutSession("Push Day", db_path=sample_db, rest_duration=10)
     start = session.current_set_start_time
 
     # simulate beginning a set 5 seconds later
@@ -143,7 +144,7 @@ def test_undo_set_start_returns_to_rest(sample_db, monkeypatch):
 
 
 def test_edit_set_overwrites_in_place(sample_db):
-    session = core.WorkoutSession("Push Day", db_path=sample_db, rest_duration=1)
+    session = WorkoutSession("Push Day", db_path=sample_db, rest_duration=1)
     session.record_metrics(session.current_exercise, session.current_set, {"Reps": 10})
     session.mark_set_completed()
     session.record_metrics(session.current_exercise, session.current_set, {"Reps": 8})
